Rename the module to `conventional_box`
