+++ image_filters/vello_filters_cpu/src/blur/conventional_box.rs
